Refactor include paths for better namespacing
