+++ lib/CException-1.3.4/CMakeLists.txt
@@ -4,6 +4,9 @@ add_library(cexception STATIC
 )
 
 target_include_directories(cexception
+    INTERFACE
+        # Include path for CException starts with "lib/CException/..."
+        ${CMAKE_CURRENT_SOURCE_DIR}/../..
     PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}
 )
