+++ src/system/CMakeLists.target.txt
@@ -16,7 +16,10 @@ target_sources(pslab-system
 )
 
 target_include_directories(pslab-system
-    PUBLIC
+    INTERFACE
+        # Include paths for pslab-system start with "system/..."
+        ${CMAKE_CURRENT_SOURCE_DIR}/..
+    PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}
 )
