+++ src/system/bus/CMakeLists.target.txt
@@ -1,12 +1,12 @@
 # Bus subsystem - target build configuration
 
+target_include_directories(pslab-system
+    PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}
+)
+
 target_sources(pslab-system
     PRIVATE
         uart.c
         usb.c
-)
-
-target_include_directories(pslab-system
-    PUBLIC
-        ${CMAKE_CURRENT_SOURCE_DIR}
 )
