+++ src/system/bus/usb.h
@@ -45,10 +45,11 @@
 #ifndef PSLAB_USB_H
 #define PSLAB_USB_H
 
-#include "util.h"
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
+
+#include "util/util.h"
 
 #ifdef __cplusplus
 extern "C" {
