+++ src/system/system.h
@@ -10,7 +10,7 @@
 #ifndef SYSTEM_H
 #define SYSTEM_H
 
-#include "fixed_point.h"
+#include "util/fixed_point.h"
 
 #define SYSTEM_VDD (FIXED_FROM_FLOAT(3.3F)) // NOLINT(readability-magic-numbers)
